Update for new gc layout
